+++ src/main/java/com/imhero/user/controller/UserController.java
@@ -1,6 +1,7 @@
 package com.imhero.user.controller;
 
 import com.imhero.config.exception.Response;
+import com.imhero.user.components.AuthenticatedUser;
 import com.imhero.user.dto.UserDto;
 import com.imhero.user.dto.request.LoginRequest;
 import com.imhero.user.dto.request.UserRequest;
@@ -11,13 +12,11 @@ import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.web.context.HttpSessionSecurityContextRepository;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
 
 @RestController
 @RequiredArgsConstructor
@@ -25,6 +24,7 @@ public class UserController {
 
     private final UserService userService;
     private final AuthenticationManager authenticationManager;
+    private final AuthenticatedUser authenticatedUser;
 
     @PostMapping("/api/v1/users")
     public Response<UserDto> join(@RequestBody UserRequest userRequest) {
@@ -32,13 +32,9 @@ public class UserController {
     }
 
     @PostMapping("/api/v1/users/login")
-    public Response<LoginResponse> login(@RequestBody LoginRequest loginRequest, HttpServletRequest request) {
+    public Response<LoginResponse> login(@RequestBody LoginRequest loginRequest) {
         Authentication authentication = authenticationManager.authenticate(new UsernamePasswordAuthenticationToken(loginRequest.getEmail(), loginRequest.getPassword()));
         SecurityContextHolder.getContext().setAuthentication(authentication);
-
-        HttpSession session = request.getSession();
-        session.setAttribute(HttpSessionSecurityContextRepository.SPRING_SECURITY_CONTEXT_KEY, SecurityContextHolder.getContext());
-
         return Response.success(userService.login(loginRequest));
     }
 }
