+++ src/main/java/com/imhero/user/service/CustomUserDetailsService.java
@@ -3,6 +3,7 @@ package com.imhero.user.service;
 import com.imhero.config.exception.ErrorCode;
 import com.imhero.config.exception.ImheroApplicationException;
 import com.imhero.user.domain.User;
+import com.imhero.user.dto.CustomUserDetails;
 import com.imhero.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -10,11 +11,11 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
+import java.io.Serializable;
 
 @Service
 @RequiredArgsConstructor
-public class CustomUserDetailsService implements UserDetailsService {
+public class CustomUserDetailsService implements UserDetailsService, Serializable {
 
     private final UserRepository userRepository;
 
@@ -23,19 +24,5 @@ public class CustomUserDetailsService implements UserDetailsService {
         User user = userRepository.findUserByEmail(email)
                 .orElseThrow(() -> new ImheroApplicationException(ErrorCode.EMAIL_NOT_FOUND));
         return new CustomUserDetails(user);
-    }
-
-    public class CustomUserDetails extends org.springframework.security.core.userdetails.User {
-
-        private User user;
-
-        private CustomUserDetails(User user) {
-            super(user.getEmail(), user.getPassword(), new ArrayList<>());
-            this.user = user;
-        }
-
-        public User getUser() {
-            return user;
-        }
     }
 }
